Proper main(), globals specified, adopted tab offset of 2
A tab offset of 2 is the Google Coding Style Guideline, and well, I
was already beginning to see so much indent that I couldn't clearly
delineate blocks at visual glance in certain cases. If you use IDLE,
you can adjust this.  Otherwise, the tab-indent-offset tag at the top
of the file instructs emacs and emacs-aware programming editors what
the indent offset style is.