Completed evt-realloc: auto-namespace all events in a full mod suite
The only feature lacking is command-line parameter support, although
it'd be trivial to add given the way the code is presently structured.

In the interim, the variables which would have been sourced from the
command-line are simply directly set at the top of the script. Presently,
they're all configured to auto-namespace all the numeric events in
the entire VIET repository, a task now complete thanks to this app.