hiphub: generate EMF beta with a fully checksummed manifest
I'm moving in the direction of the installer being strict about only
allowing installations that fully meet the checksum & manifest
information (though there will be a dev workaround) supplied with the
release. This is to stop having the extra installation correctness
hurdle of always remembering to delete the prior HIP version's
"modules/" folder. Even if the installer does already warn like hell
if the installed files do not match the released files, people don't
pay attention, and it's easy to make this mistake.

To still allow for the EMF Beta to be installed, we must also
automatically package it with its own, newer manifest pertaining to
its HIP modules. Thus, this is the hiphub side of that deal.